get_coordinates
