+++ src/app/core/services/worker_python/python-packages.json
@@ -90,7 +90,7 @@
     "source": "cdn"
   },
   "stellar-engine": {
-    "file_name": "stellar_engine-0.1.3-cp313-none-any.whl",
+    "file_name": "stellar_engine-0.1.3-py3-none-any.whl",
     "name": "stellar_engine",
     "source": "local"
   },
